Refactor data display items, tooltip
