+++ containers/web-server/frontend/src/details/regions/RegionDetailsContent.tsx
@@ -3,9 +3,9 @@ import { FC } from 'react';
 
 import { InteractionTarget } from '@/lib/data-map/interactions/use-interactions';
 import { numFormat } from '@/lib/helpers';
+import { DataItem } from '@/lib/ui/data-display/DataItem';
 
 import { REGIONS_METADATA } from '@/config/regions/metadata';
-import { DataItem } from '@/details/features/detail-components';
 
 export const RegionDetailsContent: FC<{ selectedRegion: InteractionTarget<any> }> = ({ selectedRegion }) => {
   const metadata = REGIONS_METADATA[selectedRegion.viewLayer.params.regionLevel];
