+++ containers/web-server/frontend/src/details/solutions/SolutionsSidebarContent.tsx
@@ -3,11 +3,11 @@ import _ from 'lodash';
 import { FC } from 'react';
 
 import { colorMap } from '@/lib/color-map';
+import { ColorBox } from '@/lib/ui/data-display/ColorBox';
+import { DataItem } from '@/lib/ui/data-display/DataItem';
 
 import { VECTOR_COLOR_MAPS } from '@/config/color-maps';
 import { MARINE_HABITATS_LOOKUP } from '@/config/solutions/domains';
-import { DataItem } from '@/details/features/detail-components';
-import { ColorBox } from '@/map/tooltip/content/ColorBox';
 import { habitatColorMap } from '@/state/layers/data-layers/marine';
 import { landuseColorMap } from '@/state/layers/data-layers/terrestrial';
