+++ containers/web-server/frontend/src/lib/ui/data-display/ColorBox.tsx
