+++ containers/web-server/frontend/src/map/tooltip/TooltipContent.tsx
@@ -6,12 +6,12 @@ import { hasHover, hoverState } from '@/lib/data-map/interactions/interaction-st
 import { InteractionTarget } from '@/lib/data-map/interactions/use-interactions';
 import { ErrorBoundary } from '@/lib/react/ErrorBoundary';
 
-import { DroughtHoverDescription } from './content/DroughtHoverDescription';
+import { NETWORKS_METADATA } from '@/config/networks/metadata';
+
+import { VectorHoverDescription } from './VectorHoverDescription';
 import { HazardHoverDescription } from './content/HazardHoverDescription';
 import { HdiHoverDescription } from './content/HdiHoverDescription';
 import { PopulationHoverDescription } from './content/PopulationHoverDescription';
-import { SolutionHoverDescription } from './content/SolutionHoverDescription';
-import { VectorHoverDescription } from './content/VectorHoverDescription';
 
 const TooltipSection = ({ children }) => (
   <Box px={1} py={0.5} borderBottom="1px solid #ccc">
@@ -24,17 +24,12 @@ export const TooltipContent: FC = () => {
   const hoveredHazards = useRecoilValue(hoverState('hazards')) as InteractionTarget<any>[];
   const hoveredPopulation = useRecoilValue(hoverState('population')) as InteractionTarget<any>;
   const hoveredRegion = useRecoilValue(hoverState('regions')) as InteractionTarget<any>;
-  const hoveredSolution = useRecoilValue(hoverState('solutions')) as InteractionTarget<any>;
-  const hoveredDrought = useRecoilValue(hoverState('drought')) as InteractionTarget<any>;
 
   const assetsHovered = hasHover(hoveredVector);
   const hazardsHovered = hasHover(hoveredHazards);
   const populationHovered = hasHover(hoveredPopulation);
   const regionsHovered = hasHover(hoveredRegion);
-  const solutionsHovered = hasHover(hoveredSolution);
-  const droughtHovered = hasHover(hoveredDrought);
-  const doShow =
-    assetsHovered || hazardsHovered || populationHovered || regionsHovered || solutionsHovered || droughtHovered;
+  const doShow = assetsHovered || hazardsHovered || populationHovered || regionsHovered;
 
   if (!doShow) return null;
 
@@ -42,19 +37,10 @@ export const TooltipContent: FC = () => {
     <Paper>
       <Box minWidth={200}>
         <ErrorBoundary message="There was a problem displaying the tooltip.">
-          {solutionsHovered && (
-            <TooltipSection>
-              <SolutionHoverDescription hoveredObject={hoveredSolution} />
-            </TooltipSection>
-          )}
+          {/* TODO: generate tooltip contents straight from view layers */}
           {assetsHovered ? (
             <TooltipSection>
-              <VectorHoverDescription hoveredObject={hoveredVector} />
-            </TooltipSection>
-          ) : null}
-          {droughtHovered ? (
-            <TooltipSection>
-              <DroughtHoverDescription hoveredObject={hoveredDrought} />
+              <VectorHoverDescription hoveredObject={hoveredVector} metadataLookup={NETWORKS_METADATA} />
             </TooltipSection>
           ) : null}
           {hazardsHovered ? (
